fixed actions refersher

+++ telegram/jobs/actions_refresher.py
@@ -1,5 +1,6 @@
 from datetime import datetime, timedelta
 
+from config.enums import ActionEnum
 from database.repo import Repos
 from database.session import get_db
 
@@ -10,5 +11,5 @@ async def actions_refresher():
 
         actions = await repos.action.get_all_actions()
         for action in actions:
-            if action.created_at <= datetime.now() - timedelta(days=10):
+            if action.action_type == ActionEnum.skip and action.created_at <= datetime.now() - timedelta(days=10):
                 await repos.action.delete_action(action.id)
